fixed database stuff

+++ INTEX_AURORA_BRICKS/Models/Products.cs
@@ -5,15 +5,15 @@
     public class Products
     {
         [Key]
-        public byte product_ID { get; set; }
+        public short product_ID { get; set; }
 
         public string? name { get; set; }
 
-        public short? year { get; set; }
+        public int? year { get; set; }
 
-        public short? num_parts { get; set; }
+        public int? num_parts { get; set; }
 
-        public short? price { get; set; }
+        public int? price { get; set; }
 
         public string? img_link { get; set; }
